Adds tests for message encode / decode Adds ability to encode / decode entire messages
